NEWEST UPDATE
Updated the buttons to allow for hover, updated all the buttons so that sides can no longer be clicked.
All the new button hover colors are added in according to Trixie.
Removed all the text from footers Icon.
Moved the footer on game hosting sites to be lower abit.
Added in different css files for each respective type of file, check accordingly.
Will be working on responsiveness next